Fixed backup
The backup function was causing to delete everything in the file a user has chosen in the kdialog if the user didn't create a subdirectory.
This is now fixed, the script will ask for the directory where the Shortix-Backup directory should be created and create this new directory. Same for the installer without kdialog.
The downside is that the user can't choose the Shortix-Backup directory name.

And I also fixed the backup for systems without kdialog - it should work now.